backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pyodbc>=4.0.32
pandas>=1.3.3
python-dotenv>=0.19.0
pydantic>=1.8.2
openpyxl>=3.0.9
python-multipart>=0.0.5
numpy>=1.20.0
jwt>=1.3.1
pyjwt>=2.1.0
pathlib>=1.0.1
uuid>=1.30
# Logging-related dependencies
python-json-logger>=2.0.4
